backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
starlette==0.27.0
google-auth-oauthlib==1.2.0
google-auth==2.25.2
requests==2.31.0
